+++ examples/index.html
@@ -49,6 +49,9 @@
   <body>
     <div id="top-left">
       <div class="row">
+        <a href="import.html">Create your own graph files</a>
+      </div>
+      <div class="row">
         <label
           >Change graph file:
           <input type="file" id="fileInput" />
