+++ osm-to-route-snapper/Cargo.toml
@@ -3,11 +3,26 @@ name = "osm-to-route-snapper"
 version = "0.1.0"
 edition = "2021"
 
+[lib]
+crate-type = ["cdylib", "rlib"]
+
 [dependencies]
 abstutil = { git = "https://github.com/a-b-street/abstreet" }
 bincode = "1.3.1"
-clap = { version = "4.0.32", features = ["derive"] }
 geom = { git = "https://github.com/a-b-street/abstreet" }
 osm2streets = { git = "https://github.com/a-b-street/osm2streets" }
 route-snapper-graph = { path = "../route-snapper-graph" }
 streets_reader = { git = "https://github.com/a-b-street/osm2streets" }
+
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+clap = { version = "4.0.32", features = ["derive"] }
+
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+console_error_panic_hook = "0.1.6"
+console_log = "0.2.0"
+# TODO Upstream this in abstutil crate. WASM is missing some runtime dep otherwise.
+instant = { version = "0.1.12", features = ["wasm-bindgen"] }
+log = "0.4.17"
+serde-wasm-bindgen = "0.4.5"
+wasm-bindgen = "0.2.70"
+web-sys = { version = "0.3.6", features = ["console"] }
